+++ core/vm/instructions.go
@@ -618,7 +618,7 @@ func opCreate(pc *uint64, interpreter *EVMInterpreter, scope *ScopeContext) ([]b
 	}
 
 	res, addr, returnGas, suberr := interpreter.evm.Create(scope.Contract, input, gas, bigVal)
-	// Special case the error in the op code
+	// Special case the error in the op code. TODO remove.
 	if errors.Is(suberr, vmerrs.ErrToAddrProhibitedSoft) {
 		return nil, suberr
 	}
@@ -668,7 +668,7 @@ func opCreate2(pc *uint64, interpreter *EVMInterpreter, scope *ScopeContext) ([]
 	}
 	res, addr, returnGas, suberr := interpreter.evm.Create2(scope.Contract, input, gas,
 		bigEndowment, &salt)
-	// Special case the error in the op code
+	// Special case the error in the op code. TODO remove.
 	if errors.Is(suberr, vmerrs.ErrToAddrProhibitedSoft) {
 		return nil, suberr
 	}
@@ -714,7 +714,7 @@ func opCall(pc *uint64, interpreter *EVMInterpreter, scope *ScopeContext) ([]byt
 	}
 
 	ret, returnGas, err := interpreter.evm.Call(scope.Contract, toAddr, args, gas, bigVal)
-	// Special case the error in the op code
+	// Special case the error in the op code. TODO remove.
 	if errors.Is(err, vmerrs.ErrToAddrProhibitedSoft) {
 		return nil, err
 	}
@@ -768,7 +768,7 @@ func opCallExpert(pc *uint64, interpreter *EVMInterpreter, scope *ScopeContext) 
 	}
 
 	ret, returnGas, err := interpreter.evm.CallExpert(scope.Contract, toAddr, args, gas, bigVal, coinID, bigVal2)
-	// Special case the error in the op code
+	// Special case the error in the op code. TODO remove.
 	if errors.Is(err, vmerrs.ErrToAddrProhibitedSoft) {
 		return nil, err
 	}
@@ -807,7 +807,7 @@ func opCallCode(pc *uint64, interpreter *EVMInterpreter, scope *ScopeContext) ([
 	}
 
 	ret, returnGas, err := interpreter.evm.CallCode(scope.Contract, toAddr, args, gas, bigVal)
-	// Special case the error in the op code
+	// Special case the error in the op code. TODO remove.
 	if errors.Is(err, vmerrs.ErrToAddrProhibitedSoft) {
 		return nil, err
 	}
@@ -840,7 +840,7 @@ func opDelegateCall(pc *uint64, interpreter *EVMInterpreter, scope *ScopeContext
 	args := scope.Memory.GetPtr(int64(inOffset.Uint64()), int64(inSize.Uint64()))
 
 	ret, returnGas, err := interpreter.evm.DelegateCall(scope.Contract, toAddr, args, gas)
-	// Special case the error in the op code
+	// Special case the error in the op code. TODO remove.
 	if errors.Is(err, vmerrs.ErrToAddrProhibitedSoft) {
 		return nil, err
 	}
@@ -873,7 +873,7 @@ func opStaticCall(pc *uint64, interpreter *EVMInterpreter, scope *ScopeContext) 
 	args := scope.Memory.GetPtr(int64(inOffset.Uint64()), int64(inSize.Uint64()))
 
 	ret, returnGas, err := interpreter.evm.StaticCall(scope.Contract, toAddr, args, gas)
-	// Special case the error in the op code
+	// Special case the error in the op code. TODO remove.
 	if errors.Is(err, vmerrs.ErrToAddrProhibitedSoft) {
 		return nil, err
 	}