+++ plugin/evm/export_tx.go
@@ -4,6 +4,7 @@
 package evm
 
 import (
+	"errors"
 	"fmt"
 	"math/big"
 
@@ -25,8 +26,10 @@ import (
 )
 
 var (
-	_ UnsignedAtomicTx       = &UnsignedExportTx{}
-	_ secp256k1fx.UnsignedTx = &UnsignedExportTx{}
+	_                           UnsignedAtomicTx       = &UnsignedExportTx{}
+	_                           secp256k1fx.UnsignedTx = &UnsignedExportTx{}
+	errExportNonAVAXInputBanff                         = errors.New("export input cannot contain non-AVAX in Banff")
+	errExportNonAVAXOutputBanff                        = errors.New("export output cannot contain non-AVAX in Banff")
 )
 
 // UnsignedExportTx is an unsigned ExportTx
@@ -92,6 +95,9 @@ func (utx *UnsignedExportTx) Verify(
 		if err := in.Verify(); err != nil {
 			return err
 		}
+		if rules.IsBanff && in.AssetID != ctx.AVAXAssetID {
+			return errExportNonAVAXInputBanff
+		}
 	}
 
 	for _, out := range utx.ExportedOutputs {
@@ -101,6 +107,9 @@ func (utx *UnsignedExportTx) Verify(
 		assetID := out.AssetID()
 		if assetID != ctx.AVAXAssetID && utx.DestinationChain == constants.PlatformChainID {
 			return errWrongChainID
+		}
+		if rules.IsBanff && assetID != ctx.AVAXAssetID {
+			return errExportNonAVAXOutputBanff
 		}
 	}
 	if !avax.IsSortedTransferableOutputs(utx.ExportedOutputs, Codec) {
@@ -158,7 +167,7 @@ func (utx *UnsignedExportTx) Burned(assetID ids.ID) (uint64, error) {
 		}
 	}
 
-	return math.Sub64(input, spent)
+	return math.Sub(input, spent)
 }
 
 // SemanticVerify this transaction is valid.