+++ tests/init.go
@@ -224,7 +224,7 @@ var Forks = map[string]*params.ChainConfig{
 		ApricotPhase4BlockTimestamp: big.NewInt(0),
 		ApricotPhase5BlockTimestamp: big.NewInt(0),
 	},
-	"Blueberry": {
+	"Banff": {
 		ChainID:                     big.NewInt(1),
 		HomesteadBlock:              big.NewInt(0),
 		EIP150Block:                 big.NewInt(0),
@@ -239,7 +239,25 @@ var Forks = map[string]*params.ChainConfig{
 		ApricotPhase3BlockTimestamp: big.NewInt(0),
 		ApricotPhase4BlockTimestamp: big.NewInt(0),
 		ApricotPhase5BlockTimestamp: big.NewInt(0),
-		BlueberryBlockTimestamp:     big.NewInt(0),
+		BanffBlockTimestamp:         big.NewInt(0),
+	},
+	"Cortina": {
+		ChainID:                     big.NewInt(1),
+		HomesteadBlock:              big.NewInt(0),
+		EIP150Block:                 big.NewInt(0),
+		EIP155Block:                 big.NewInt(0),
+		EIP158Block:                 big.NewInt(0),
+		ByzantiumBlock:              big.NewInt(0),
+		ConstantinopleBlock:         big.NewInt(0),
+		PetersburgBlock:             big.NewInt(0),
+		IstanbulBlock:               big.NewInt(0),
+		ApricotPhase1BlockTimestamp: big.NewInt(0),
+		ApricotPhase2BlockTimestamp: big.NewInt(0),
+		ApricotPhase3BlockTimestamp: big.NewInt(0),
+		ApricotPhase4BlockTimestamp: big.NewInt(0),
+		ApricotPhase5BlockTimestamp: big.NewInt(0),
+		BanffBlockTimestamp:         big.NewInt(0),
+		CortinaBlockTimestamp:       big.NewInt(0),
 	},
 }