+++ RELEASES.md
@@ -1,5 +1,23 @@
 # Release Notes
 
+## [v0.11.0](https://github.com/ava-labs/coreth/releases/tag/v0.10.0)
+
+- Update Chain Config compatibility check to compare against last accepted block timestamp
+- Bump go-ethereum dependency to v1.10.25
+- Add Banff activation times for Mainnet and Fuji for October 18 4pm UTC and October 3 2pm UTC respectively
+- Banff cleanup
+
+## [v0.10.0](https://github.com/ava-labs/coreth/releases/tag/v0.10.0)
+
+- Deprecate Native Asset Call and Native Asset Balance
+- Deprecate Import/Export of non-AVAX Avalanche Native Tokens via Atomic Transactions
+- Add failure reason to bad block API
+
+## [v0.9.0](https://github.com/ava-labs/coreth/releases/tag/v0.9.0)
+
+- Migrate to go-ethereum v1.10.23
+- Add API to fetch Chain Config
+
 ## [v0.8.16](https://github.com/ava-labs/coreth/releases/tag/v0.8.16)
 
 - Fix bug in `codeToFetch` database accessors that caused an error when starting/stopping state sync